feat: Add loading message in callback function
Add a loading message in the callback function of the app.py file to display a loading animation in the Line messaging platform. This is achieved by calling the `loading_message` method from the `LINE_API_UTIL` class in the line_api_util.py file. The loading message is sent to the user identified by their user ID.