feat: Add view info action in contact and course bots
Adds support in the contact bot to handle 'view info' payloads and display
individual contact information.

In the course bot, adds an action to view teacher information when there is only
one teacher. Limits the number of teacher actions to two when there are multiple
teachers to prevent UI overload.